portfolioProject/requirements.txt
anyio==3.3.4
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
config==0.5.1
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.47.0
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-colorfield==0.5.0
django-config==0.1.4
django-crispy-forms==1.13.0
django-decouple==2.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-restframework==0.0.1
django-six==1.0.4
django-social-share==2.2.1
django-utils-six==2.0
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
HTMLTestReport==1.0.1
httpcore==0.13.7
httpx==0.20.0
idna==3.3
Markdown==3.3.4
oauthlib==3.1.1
outcome==1.1.0
Pillow==8.4.0
psycopg2==2.9.3
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.1
whitenoise==5.3.0
wsproto==1.0.0
youtube-search-python==1.5.1
